+++ qsprpred/data/utils/descriptorsets.py
@@ -6,11 +6,14 @@ To add a new descriptor or fingerprint calculator:
 """
 
 from abc import ABC, abstractmethod
-from typing import Union, List
+from typing import Optional, Union, List
 
 import mordred
 import numpy as np
 from mordred import descriptors as mordreddescriptors
+from Mold2_pywrapper import Mold2 as Mold2_calculator
+from PaDEL_pywrapper import PaDEL as PaDEL_calculator
+from PaDEL_pywrapper.descriptors import descriptors as PaDEL_descriptors
 from qsprpred.data.utils.descriptor_utils.drugexproperties import Property
 from qsprpred.data.utils.descriptor_utils.rdkitdescriptors import RDKit_desc
 from rdkit import Chem, DataStructs
@@ -35,7 +38,7 @@ class DescriptorSet(ABC):
 
     def iterMols(self, mols : List[Union[str, Mol]], to_list=False):
         """
-        Calculate the descriptor for a molecule.
+        Create a molecule iterator or list from RDKit molecules or SMILES.
 
         Args:
             mols: list of molecules (SMILES `str` or RDKit Mol)
@@ -150,7 +153,7 @@ class Mordred(DescriptorSet):
 
     From https://github.com/mordred-descriptor/mordred.
     Initialize the descriptor with the same arguments as you would pass to `Calculator` function of Mordred.
-    If no mordred descriptor object passed, the all descriptors will be calculated.
+    If no mordred descriptor object passed, then all descriptors are calculated.
 
     Args:
         *args: `Calculator` arguments
@@ -205,7 +208,7 @@ class Mordred(DescriptorSet):
 
 class DrugExPhyschem(DescriptorSet):
     """
-    Pysciochemical properties originally used in DrugEx for QSAR modelling
+    Physicochemical properties originally used in DrugEx for QSAR modelling
     Initialize the descriptor with Property arguments (a list of properties to calculate) to select a subset.
 
     Args:
@@ -251,8 +254,8 @@ class rdkit_descs(DescriptorSet):
     Add compute_3Drdkit argument to indicate if 3D descriptors should also be calculated
 
     Args:
-        *args: `Property` arguments
-        **kwargs: `Property` keyword arguments
+        *args: `RDKit_desc` arguments
+        **kwargs: `RDKit_desc` keyword arguments
     """
 
     def __init__(self, *args, **kwargs):
@@ -284,6 +287,130 @@ class rdkit_descs(DescriptorSet):
 
     def __str__(self):
         return "RDkit"
+
+
+class Mold2(DescriptorSet):
+    """Descriptors from molecular descriptor calculation software Mold2.
+
+    From https://github.com/OlivierBeq/Mold2_pywrapper.
+    Initialize the descriptor with no arguments.
+    All descriptors are always calculated.
+    """
+
+    def __init__(self, *args, **kwargs):
+        """Initialize a PaDEL calculator
+
+        Args:
+            *args: ignored
+            **kwargs: ignored
+        """
+        # args and kwargs are ignored
+        self._args = None
+        self._kwargs = None
+
+        self._is_fp = False
+
+        self._mold2 = Mold2_calculator()
+        self._descriptors = self._mold2.calculate([Chem.MolFromSmiles("C")], show_banner=False).columns.tolist()
+        self._keepindices = list(range(len(self._descriptors)))
+
+    def __call__(self, mols):
+        values = self._mold2.calculate(self.iterMols(mols), show_banner=False, njobs=1).values
+        # Drop columns
+        return values[:, self._keepindices]
+
+    @property
+    def is_fp(self):
+        return self._is_fp
+
+    @property
+    def settings(self):
+        return self._args, self._kwargs
+
+    @property
+    def descriptors(self):
+        return self._descriptors
+
+    @descriptors.setter
+    def descriptors(self, names):
+        # Remove names not listed
+        for i, index, desc_name in zip(range(-1, -len(self._keepindices)-1, -1),
+                                       self._keepindices[::-1],
+                                       self.descriptors[::-1]):
+            if desc_name not in names:
+                del self._keepindices[i], self._descriptors[i]
+
+    def __str__(self):
+        return "Mold2"
+
+
+class PaDEL(DescriptorSet):
+    """Descriptors from molecular descriptor calculation software PaDEL.
+
+    From https://github.com/OlivierBeq/PaDEL_pywrapper.
+    """
+
+    def __init__(self, descs: Optional[List[str]] = None, **kwargs):
+        """Initialize a PaDEL calculator
+
+        Args:
+            descs: list of PaDEL descriptor short names
+            **kwargs: keyword arguments of the `PaDEL_pywrapper.PaDEL` constructor
+        """
+        # args and kwargs are ignored
+        self._args = descs
+        self._process_kwargs(**kwargs)
+
+        self._is_fp = False
+
+        # Create calculator and obtain default descriptor names
+        dummy_mol = Chem.AddHs(Chem.MolFromSmiles("CC"))
+        AllChem.EmbedMolecule(dummy_mol)  # Required for line below, since 3D descs would raise without 3D coords
+        self._name_mapping = {name: descriptor
+                              for descriptor in PaDEL_descriptors
+                              for name in descriptor().calculate(dummy_mol)}
+
+        # Initialize descriptor names
+        if descs is None:
+            self._descriptors = PaDEL_descriptors
+        else:
+            self.descriptors = descs
+        # Instantiate calculator
+        self._padel = PaDEL_calculator(self._descriptors, ignore_3D=self._ignore_3D)
+        # Obtain filtered descriptors from calculator
+        self._descriptors = self._padel.descriptors
+
+    def __call__(self, mols):
+        values = self._padel.calculate(list(self.iterMols(mols)), show_banner=False, njobs=1).values
+        return values
+
+    @property
+    def is_fp(self):
+        return self._is_fp
+
+    @property
+    def settings(self):
+        return self._args, self._kwargs
+
+    @property
+    def descriptors(self):
+        names = [name for name, desc in self._name_mapping.items() if desc in self._descriptors]
+        return names
+
+    @descriptors.setter
+    def descriptors(self, names):
+        self._descriptors = [self._name_mapping[name] for name in names]
+        self._padel = PaDEL_calculator(self._descriptors, ignore_3D=self._ignore_3D)
+
+    def get_names(self):
+        return self._padel.calculate([Chem.MolFromSmiles("C")], show_banner=False).columns.tolist()
+
+    def _process_kwargs(self, **kwargs):
+        self._ignore_3D = kwargs.get("ignore_3D", True)
+        self._kwargs = kwargs
+
+    def __str__(self):
+        return "PaDEL"
 
 
 class _DescriptorSetRetriever: